✨ feat: Add image and update README
This commit adds an image to the README.md file and updates the content. The image `terraform-rancher2.png` is included in the assets folder. Additionally, the README.md file now contains instructions on how to initialize rancher 2 with terraform. Users are also encouraged to star the repository if they find it useful.